[message.rb]: use helper to distinguish between webhook responses
Adds two helper functions that encapsulate the properties of a
successful connector based webhook and a workflow based webhook.

Up to now, workflow based webhooks lead to a response where there
response code is 202 and the response body is empty.
For connector based webhooks, the status code is 200 and the response
body is set to "1".