Support msteams workflows (#10)
* [message.rb]: response.body can be empty on workflow responses

* [complex_example.rb]: Use dynamic datetime in text

* [message.rb]: use helper to distinguish between webhook responses

Adds two helper functions that encapsulate the properties of a
successful connector based webhook and a workflow based webhook.

Up to now, workflow based webhooks lead to a response where there
response code is 202 and the response body is empty.
For connector based webhooks, the status code is 200 and the response
body is set to "1".

* [complex_example.rb]: Rubocop says to use '.to_s`

* [message.rb]: Add comment and use '.' to call functions

* Move funcs to identify response type into class Message

as per PR review

* Check whether sent msg is too large in dedicated method

- Minor formatting change to simplify 'return response' statement

* [message.rb] Use '&&' instead of 'and' more ruby-ish?